requirements.txt
flask==3.1.0
flask-cors==5.0.1
yfinance==0.2.54
pandas==2.2.3
pandas-ta==0.3.14b1
numpy==1.26.4
python-telegram-bot==20.8
